requirements.txt
absl-py==1.4.0
astunparse==1.6.3
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
geotorch==0.3.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.13.4
idna==2.10
imageio==2.28.1
importlib-metadata==6.3.0
importlib-resources==5.12.0
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.25.1
requests-oauthlib==1.3.1
robustbench @ git+https://github.com/RobustBench/robustbench.git@dc7016009f62c93f568a0e29aa3413af175f9ae5
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tifffile==2023.4.12
timm==0.6.13
torch==2.0.0
torchaudio==2.0.1
torchdiffeq==0.2.3
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
